% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChMarkov.R
\name{ChMarkov}
\alias{ChMarkov}
\title{Simulation des chaines de Markov (TP N02)}
\usage{
ChMarkov(etats, mu, p, nb_rep)
}
\arguments{
\item{etats}{Le vecteur des etats de la chaine}

\item{mu}{La disttibution intiale}

\item{p}{La matrice de trasition}

\item{nb_rep}{Le nom de repetition des trasitions}
}
\description{
Simulation d'une chaine de Markov
}
