% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdist.R
\name{rdist}
\alias{rdist}
\title{Simulation d'une loi discrete}
\usage{
rdist(x, p)
}
\arguments{
\item{x}{vecteur numerique  variable aleatoire}

\item{p}{vecteur numerique  variable de propabilite}
}
\description{
Simulation d'une loi discrete
}
